Redesign the application
for this contact app, i built two view controller with navigation one, first view controller is the lunching one that contains Table view with its lonely cell that contains circled and grey bordered image of the person and label for its name and another label for its status. I designed the cell so the user could press to move to secondViewController for the next page of the application, which contains his circled image with his name, furthermore his name again with his names,numbers, status and images are in arrays.