user default and textfield delegate function are working
adding data to user default is done correctly but doesn't appear in the table view cell after relaunching the application that was due to contacts.count in tabelviewcells number function = 0, the solution was to assign array returned from retrievedata function to contacts array in init method in ContactListViewModel so it is assigned every time we deal with this viewmodel, we reach a point in which we deal with data from one place ViewModel by set to and get from contacts array which secretly retrieve data from retrieve function, another one that textfield doesn't work as expected that was because of not assigning delegate to that textfield in cell when it has number label so we  have to tell function to work in this specific cell by assigning userTextfield.delegate=self when rowType that i got from addContactViewController in dequeue cell function is equal to number, all this work handled in didSet to do it every time rowType change. I handled one of the deprecated method to work with iOS15.