NCContacts Store
I edit the code to grab data from iPhone's contacts rather than core data by using NCContacts as the following steps:
1- Get access to the contacts store by making instance of CNContactStore().
2-Specify which data keys we want to fetch[CNContactGivenNameKey, CNContactPhoneNumbersKey, CNContactImageDataKey].
3- Create fetch request based on the key chosen in step 2.
4- Call method to fetch all contacts(iterate over all contacts by storing values in person object then append object to contacts array in the closure (enumerateContacts)). Moreover, I give the user ability to allow application access contacts by adding Privacy - Contacts Usage Description in info tab from project settings. then calling fetch function from init function but important to notice that it must be run in the background so it can't break mian thread by using  DispatchQueue.global(qos: .background).asyn then I reload table view by implementing delegation pattern as a result we achieved abstraction concept of OOP.
Delegation pattern was to make a connection between ContactsListViewModel(CLVM) and ContactsListViewController(CLVC),  in CLVM I make protocol RefreshDelegate which is a map of what function i can use in CLVC, then i defined delegate variable of type RefreshDelegate in CLVM to bind ContactListViewModel to ContactListViewController. in CLVC, I make an extension of RefreshDelegate and implement Refresh function inside it which responsible to reload table view model in the main thread with the new data from contacts feature in iphone.
At last, i change the name of protocol to be clear from RefreshDelegate to ContactsListViewModelDelegate.